split built-in functions to libs
