+++ core/schema.go
@@ -1,4 +1,4 @@
-package yisp
+package core
 
 import (
 	"encoding/json"
