+++ hoge.yaml
@@ -1,0 +1,9 @@
+!yisp
+- crypto.sha256
+- - yaml.marshal
+  - hoge: piyo
+    a: 1
+    b: 2
+    c:
+      d: 3
+      e: 4
